Add more complete config for Bluemix and add info to readme
Will properly tag/push to Bluemix and run on the container service if
users follow the configuration/docs in the README.md

Also made sure that the docker images do not contain the dot-env or
.env files so that no secrets/credentials are in the image itself.
Future examples with Kubernetes and Swarm will show the use of secrets
as an improvement over the env. variable injection into `docker run`.

Signed-off-by: Phil Estes <estesp@gmail.com>