Add rest of conversation API demo code
This adds the rest of the code used to demonstration the Watson
conversation service addition to FlightAssist at DockerCon 2017. There
is a large "FIXME" comment regarding a necessary code update to remove
the hardcoded demo of a specific flight path to "tell the story" we used
in our Austin demo.

Signed-off-by: Phil Estes <estesp@gmail.com>